smaller FFT sizes for F4

+++ hs/SynthTools/Num.hs
@@ -312,7 +312,12 @@ instance Num F3 where
 -- inefficient implementation.  It is possible to use it for smaller
 -- DFTs though: any subcycle of 65536 should work.
 
-data F4 = F4 { unF4 :: Int32 } deriving (Eq)
+-- Note that Int32 is not enough to represent the multiplication.  It
+-- could still be used as a number representation but the
+-- multiplication module operation needs to be performed with higher
+-- precision.
+
+data F4 = F4 { unF4 :: Int64 } deriving (Eq)
 instance Show F4 where show (F4 n) = show n
 
 modF4 = nnMod 65537
